Add ranges of  ints, longs, chars. Add endingAt as companion to until (inclusive). Change boundaries in Sequences to longs. Add reverse() and shuffle(). Document methods in Sequence.

+++ src/main/java/org/d2ab/iterator/LimitingIterator.java
@@ -20,10 +20,10 @@ import java.util.NoSuchElementException;
 
 public class LimitingIterator<T> implements Iterator<T> {
 	private final Iterator<T> iterator;
-	private final int limit;
+	private final long limit;
 	int count;
 
-	public LimitingIterator(Iterator<T> iterator, int limit) {
+	public LimitingIterator(Iterator<T> iterator, long limit) {
 		this.iterator = iterator;
 		this.limit = limit;
 	}
